pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "external-resources"
dynamic = ["version"]
description = "Fetch and deploy external resources (CSS, JS, Fonts) for web services"
keywords = ["web", "css", "js", "fonts", "resources", "installation"]
readme = "README.rst"
requires-python = ">= 3.10"
authors = [
  {name = "Detlef Lannert", email = "detlef@lannert.de"},
]
maintainers = [
  {name = "Detlef Lannert", email = "detlef@lannert.de"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Django",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
  "Topic :: Software Development :: Build Tools",
  "Typing :: Typed",
]

dependencies = [
  "attrs",
  "cattrs",
  "cyclopts",
  "httpx",
  "pyyaml",
  "packaging",
  "pyproject-parser",
  "rich",
]

[project.optional-dependencies]
dev = [
  "build",
  "ipython",
  "invoke",
  "mypy",
  "pip",
  "pyproject-parser[cli]",
  "pytest",
  "types-invoke",
  "types-PyYAML",
  "wheel",
]

[project.scripts]
extres = "external_resources.extres:app"

[project.urls]
#Homepage = "https://example.com"
#Documentation = "https://readthedocs.org"
#Repository = "https://github.com/me/spam.git"
#Issues = "https://github.com/me/spam/issues"
#Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

# [tool.ext_resources]
# resource_index = "external_resources.yaml"
# requires = [
#   "bootstrap",
#   "htmx == 1.9.10",
# ]

[tool.hatch.version]
path = "src/external_resources/VERSION"
pattern = "(?P<version>\\S+)"

[tool.ruff]
ignore = ["F401", "F405"]
line-length = 100
target-version = "py310"
task-tags = ["XXX"]
extend-exclude = ["migrations"]

[tool.ruff.per-file-ignores]
# "filename" = [list of codes]

[tool.uv]
package = true
